refactor(api): place authentication routes in one file
