+++ api/server.js
@@ -2,8 +2,7 @@ import cors from "cors";
 import express from "express";
 import path from "path";
 
-import { loginRoute } from "./src/authentication/routes/login-route.js";
-import { registerRoute } from "./src/authentication/routes/register-route.js";
+import authentication from "./src/authentication/routes.js";
 import { healthCheckRoute } from "./src/shared/routes/health-check-route.js";
 
 const server = express();
@@ -12,7 +11,6 @@ server.use(cors());
 server.use(express.json());
 server.use(express.static(path.join(process.cwd(), "dist")));
 server.use(healthCheckRoute);
-server.use(registerRoute);
-server.use(loginRoute);
+server.use(authentication);
 
 export default server;
